Adding in all user endpoint logic
also a bunch of general changes to the models
to allow for better cross referencing.

and some ease of life changes for debugging to the
index.html for easy ways to test the endpoints,
and so it always uses the url it is hosted at rather
than need a url set.